# WIP: Add an iterator implementation for String splitting