Per-request of #20603 and also addresses #20628. Simple timing benchmarks show that using and collecting from these iterators are slower and cause more memory allocation.

Here's a sample timing after warming up the JIT:

```julia
julia> str = randstring(1_000_000);

# without iterator
julia> @time split(str, "a");
  0.005106 seconds (16.20 k allocations: 762.641 KB)

# with iterator
julia> @time collect(split(str, "a"));
  0.012568 seconds (194.29 k allocations: 3.957 MiB)
```

For a small string:

```julia
julia> small_str = "how-now-brown-cow";

# without iterator
julia> @time collect(split(small_str, "-"));
  0.000012 seconds (12 allocations: 544 bytes)

# with iterator
julia> @time collect(split(small_str, "-"));
  0.000048 seconds (20 allocations: 768 bytes)
```

This performance issue also seems to afflict the `GraphemeIterator` which is constructed by `graphemes(::AbstractString)` which can be optimized by implementing it as `split(str, "")` rather than returning an iterator.

Some tests are currently broken since downstream uses of `split` expect an array-like value where `length` and `getindex` can be called.

Looking for feedback before fixing these tests to see whether the maintainers-at-large still think that this abstraction is worth the performance cost or whether the unexpected allocations should be investigated.